+++ lib/wtr.ml
@@ -40,16 +40,16 @@ type 'a decoder =
   ; decode: string -> 'a option
   ; id: 'a id }
 
-let create_decoder ~name ~decode =
+let decoder ~name ~decode =
   let id = new_id () in
   {name; decode; id}
 
-let int = create_decoder ~name:"int" ~decode:int_of_string_opt
-let int32 = create_decoder ~name:"int32" ~decode:Int32.of_string_opt
-let int64 = create_decoder ~name:"int64" ~decode:Int64.of_string_opt
-let float = create_decoder ~name:"float" ~decode:float_of_string_opt
-let string = create_decoder ~name:"string" ~decode:(fun a -> Some a)
-let bool = create_decoder ~name:"bool" ~decode:bool_of_string_opt
+let int = decoder ~name:"int" ~decode:int_of_string_opt
+let int32 = decoder ~name:"int32" ~decode:Int32.of_string_opt
+let int64 = decoder ~name:"int64" ~decode:Int64.of_string_opt
+let float = decoder ~name:"float" ~decode:float_of_string_opt
+let string = decoder ~name:"string" ~decode:(fun a -> Some a)
+let bool = decoder ~name:"bool" ~decode:bool_of_string_opt
 
 type method' =
   [ `GET
@@ -64,6 +64,18 @@ type method' =
 
 let method_equal (meth1 : method') (meth2 : method') = compare meth1 meth2 = 0
 
+let method' meth =
+  match String.uppercase_ascii meth with
+  | "GET" -> `GET
+  | "HEAD" -> `HEAD
+  | "POST" -> `POST
+  | "PUT" -> `PUT
+  | "DELETE" -> `DELETE
+  | "CONNECT" -> `CONNECT
+  | "OPTIONS" -> `OPTIONS
+  | "TRACE" -> `TRACE
+  | header -> `Method header
+
 let pp_method fmt t =
   ( match t with
   | `GET -> "GET"
@@ -77,91 +89,105 @@ let pp_method fmt t =
   | `Method s -> Format.sprintf "Method (%s)" s )
   |> Format.fprintf fmt "%s"
 
-let method' meth =
-  match String.uppercase_ascii meth with
-  | "GET" -> `GET
-  | "HEAD" -> `HEAD
-  | "POST" -> `POST
-  | "PUT" -> `PUT
-  | "DELETE" -> `DELETE
-  | "CONNECT" -> `CONNECT
-  | "OPTIONS" -> `OPTIONS
-  | "TRACE" -> `TRACE
-  | header -> `Method header
-
-(** [('a, 'b) uri] represents a uniform resource identifier. The variant members
-    describe the uri component types.
-
-    - Literal Uri path literal string component eg. 'home' in '/home'
-    - Decoder Uri path argument component, i.e. the value is determined during
-      runtime, eg. ':int' in '/home/:int' *)
 type ('a, 'b) uri =
   | Nil : ('b, 'b) uri
-  | Method : method' * ('a, 'b) uri -> ('a, 'b) uri
-  | Full_splat : (string -> 'b, 'b) uri
+  | Splat : (string -> 'b, 'b) uri
   | Trailing_slash : ('b, 'b) uri
   | Literal : string * ('a, 'b) uri -> ('a, 'b) uri
-  | Decoder : 'c decoder * ('a, 'b) uri -> ('c -> 'a, 'b) uri
+  | Query_literal : string * string * ('a, 'b) uri -> ('a, 'b) uri
+  | Decode : 'c decoder * ('a, 'b) uri -> ('c -> 'a, 'b) uri
+  | Query_decode : string * 'c decoder * ('a, 'b) uri -> ('c -> 'a, 'b) uri
 
 let rec pp_uri : type a b. Format.formatter -> (a, b) uri -> unit =
- fun fmt -> function
+ fun fmt uri ->
+  let pp_query_toks = ref false in
+  match uri with
   | Nil -> Format.fprintf fmt "%!"
-  | Method (meth, uri) -> Format.fprintf fmt "%a%a" pp_method meth pp_uri uri
-  | Full_splat -> Format.fprintf fmt "/**%!"
+  | Splat -> Format.fprintf fmt "/**%!"
   | Trailing_slash -> Format.fprintf fmt "/%!"
   | Literal (lit, uri) -> Format.fprintf fmt "/%s%a" lit pp_uri uri
-  | Decoder (decoder, uri) ->
-      Format.fprintf fmt "/:%s%a" decoder.name pp_uri uri
+  | Query_literal (name, value, uri) ->
+      let pp' fmt uri = Format.fprintf fmt "%s=%s%a" name value pp_uri uri in
+      if not !pp_query_toks then (
+        pp_query_toks := true ;
+        Format.fprintf fmt "?%a" pp' uri )
+      else Format.fprintf fmt "&%a" pp' uri
+  | Decode (decoder, uri) -> Format.fprintf fmt "/:%s%a" decoder.name pp_uri uri
+  | Query_decode (name, decoder, uri) ->
+      let pp' fmt uri =
+        Format.fprintf fmt "%s=:%s%a" name decoder.name pp_uri uri
+      in
+      if not !pp_query_toks then (
+        pp_query_toks := true ;
+        Format.fprintf fmt "?%a" pp' uri )
+      else Format.fprintf fmt "&%a" pp' uri
 
-type 'c route = Route : ('a, 'c) uri * 'a -> 'c route
+type 'c route = Route : method' * ('a, 'c) uri * 'a -> 'c route
 
-let route : ('a, 'b) uri list -> 'a -> 'b route list =
- fun uris f -> List.map (fun uri' -> Route (uri', f)) uris
+let route : ?method':method' -> ('a, 'b) uri -> 'a -> 'b route =
+ fun ?(method' = `GET) uri f -> Route (method', uri, f)
+
+let routes methods uri f =
+  List.map (fun method' -> route ~method' uri f) methods
 
 let pp_route : Format.formatter -> 'b route -> unit =
- fun fmt (Route (uri, _)) -> pp_uri fmt uri
+ fun fmt (Route (method', uri, _)) ->
+  Format.fprintf fmt "%a%a" pp_method method' pp_uri uri
 
 (** Existential to encode uri component/node type. *)
 type node_type =
-  | PTrailing_slash : node_type
-  | PFull_splat : node_type
-  | PLiteral : string -> node_type
-  | PMethod : method' -> node_type
-  | PDecoder : 'c decoder -> node_type
+  | NTrailing_slash : node_type
+  | NSplat : node_type
+  | NLiteral : string -> node_type
+  | NQuery_literal : string * string -> node_type
+  | NMethod : method' -> node_type
+  | NDecoder : 'c decoder -> node_type
+  | NQuery_decoder : string * 'c decoder -> node_type
 
 let node_type_equal a b =
   match (a, b) with
-  | PTrailing_slash, PTrailing_slash -> true
-  | PFull_splat, PFull_splat -> true
-  | PLiteral lit', PLiteral lit -> String.equal lit lit'
-  | PMethod meth1, PMethod meth2 -> method_equal meth1 meth2
-  | PDecoder decoder, PDecoder decoder' -> (
+  | NTrailing_slash, NTrailing_slash -> true
+  | NSplat, NSplat -> true
+  | NLiteral lit1, NLiteral lit2 -> String.equal lit2 lit1
+  | NQuery_literal (name1, value1), NQuery_literal (name2, value2) ->
+      String.equal name1 name2 && String.equal value1 value2
+  | NMethod meth1, NMethod meth2 -> method_equal meth1 meth2
+  | NDecoder decoder, NDecoder decoder' -> (
     match eq decoder'.id decoder.id with Some Eq -> true | None -> false )
+  | NQuery_decoder (name1, decoder1), NQuery_decoder (name2, decoder2) -> (
+      String.equal name1 name2
+      && match eq decoder1.id decoder2.id with Some Eq -> true | None -> false )
   | _ -> false
 
-(* [node_type_of_uri uri] converts [uri] to [node_type list]. This is done to get around OCaml
+(* [uri'_of_uri uri] converts [uri] to [node_type list]. This is done to get around OCaml
    type inference issue when using [uri] type in the [node] function below. *)
 let rec node_type_of_uri : type a b. (a, b) uri -> node_type list = function
   | Nil -> []
-  | Method (meth, uri) -> PMethod meth :: node_type_of_uri uri
-  | Trailing_slash -> [PTrailing_slash]
-  | Full_splat -> [PFull_splat]
-  | Literal (lit, uri) -> PLiteral lit :: node_type_of_uri uri
-  | Decoder (decoder, uri) -> PDecoder decoder :: node_type_of_uri uri
+  | Trailing_slash -> [NTrailing_slash]
+  | Splat -> [NSplat]
+  | Literal (lit, uri) -> NLiteral lit :: node_type_of_uri uri
+  | Query_literal (name, value, uri) ->
+      NQuery_literal (name, value) :: node_type_of_uri uri
+  | Decode (decoder, uri) -> NDecoder decoder :: node_type_of_uri uri
+  | Query_decode (name, decoder, uri) ->
+      NQuery_decoder (name, decoder) :: node_type_of_uri uri
 
-let node_type_to_string node_type =
+let pp_node_type fmt node_type =
   match node_type with
-  | PFull_splat -> Format.sprintf "/**%!"
-  | PTrailing_slash -> Format.sprintf "/%!"
-  | PLiteral lit -> Format.sprintf "/%s" lit
-  | PDecoder decoder -> Format.sprintf "/:%s" decoder.name
-  | PMethod method' -> Format.asprintf "%a" pp_method method'
+  | NSplat -> Format.fprintf fmt "/**"
+  | NTrailing_slash -> Format.fprintf fmt "/"
+  | NLiteral lit -> Format.fprintf fmt "/%s" lit
+  | NQuery_literal (name, value) -> Format.fprintf fmt "%s=%s" name value
+  | NDecoder decoder -> Format.fprintf fmt "/:%s" decoder.name
+  | NQuery_decoder (name, decoder) ->
+      Format.fprintf fmt "%s=:%s" name decoder.name
+  | NMethod method' -> Format.fprintf fmt "%a" pp_method method'
 
 (** ['a t] is a node in a trie based router. *)
 type 'a node = {route: 'a route option; node_types: (node_type * 'a node) list}
 
 let rec node : 'a node -> 'a route -> 'a node =
- fun node' (Route (uri, _) as route) ->
+ fun node' (Route (method', uri, _) as route) ->
   let rec loop node node_types =
     match node_types with
     | [] -> {node with route= Some route}
@@ -184,7 +210,7 @@ let rec node : 'a node -> 'a route -> 'a node =
             | None -> (node_type, loop empty_node node_types) :: node.node_types
             ) }
   in
-  let node_types = node_type_of_uri uri in
+  let node_types = NMethod method' :: node_type_of_uri uri in
   loop node' node_types
 
 and empty_node : 'a node = {route= None; node_types= []}
@@ -193,7 +219,7 @@ and empty_node : 'a node = {route= None; node_types= []}
    that iterating nodes via array is faster than via the list. *)
 type 'a t = {route: 'a route option; node_types: (node_type * 'a t) array}
 
-let rec create routes =
+let rec t routes =
   List.concat routes |> List.fold_left node empty_node |> compile
 
 and compile : 'a node -> 'a t =
@@ -207,29 +233,41 @@ and compile : 'a node -> 'a t =
 let rec pp fmt t =
   let nodes = t.node_types |> Array.to_list in
   let len = List.length nodes in
-  nodes
-  |> Format.pp_print_list
-       ~pp_sep:(if len > 1 then Format.pp_force_newline else fun _ () -> ())
-       (fun fmt (nt, t') ->
-         Format.pp_open_vbox fmt 2 ;
-         Format.pp_print_string fmt (node_type_to_string nt) ;
-         if Array.length t'.node_types > 0 then (
-           Format.pp_print_break fmt 0 0 ;
-           pp fmt t' ) ;
-         Format.pp_close_box fmt () ;
-         () )
-       fmt
+  let query_tok_printed = ref false in
+  Format.pp_print_list
+    ~pp_sep:(if len > 1 then Format.pp_force_newline else fun _ () -> ())
+    (fun fmt (node_type, t') ->
+      Format.pp_open_vbox fmt 2 ;
+      ( match node_type with
+      | NQuery_literal _ | NQuery_decoder _ ->
+          if not !query_tok_printed then
+            Format.fprintf fmt "?%a" pp_node_type node_type
+          else Format.fprintf fmt "&%a" pp_node_type node_type
+      | node -> Format.fprintf fmt "%a" pp_node_type node ) ;
+      if Array.length t'.node_types > 0 then (
+        Format.pp_print_break fmt 0 0 ;
+        pp fmt t' ) ;
+      Format.pp_close_box fmt () )
+    fmt nodes
 
 type decoded_value = D : 'c decoder * 'c -> decoded_value
 
-let rec match' method' uri (t : 'a t) =
+let rec drop : 'a list -> int -> 'a list =
+ fun l n -> match l with _ :: tl when n > 0 -> drop tl (n - 1) | t -> t
+
+let rec match' : method' -> string -> 'a t -> 'a option =
+ fun method' uri t ->
   (* split uri path and query into tokens *)
   let uri' = uri |> String.trim |> Uri.of_string in
-  let path_tokens = Uri.path uri' |> String.split_on_char '/' |> List.tl in
+  let path_tokens =
+    Uri.path uri'
+    |> String.split_on_char '/'
+    |> List.tl
+    |> List.map (fun tok -> `Path tok)
+  in
   let query_tokens =
     Uri.query uri'
-    |> List.map (fun (k, v) ->
-           if List.length v > 0 then [k; List.hd v] else [k] )
+    |> List.map (fun (k, values) -> List.map (fun v' -> `Query (k, v')) values)
     |> List.concat
   in
   let uri_toks = path_tokens @ query_tokens in
@@ -242,7 +280,7 @@ let rec match' method' uri (t : 'a t) =
     match uri_toks with
     | [] ->
         Option.map
-          (fun (Route (uri, f)) ->
+          (fun (Route (_, uri, f)) ->
             exec_route_handler f (uri, List.rev decoded_values) )
           t.route
     | uri_part :: uris ->
@@ -251,22 +289,24 @@ let rec match' method' uri (t : 'a t) =
         let matched_node = ref None in
         let full_splat_matched = ref false in
         while !continue && !index < Array.length t.node_types do
-          match t.node_types.(!index) with
-          | PDecoder decoder, t' -> (
-            match decoder.decode uri_part with
+          match (uri_part, t.node_types.(!index)) with
+          | `Path v, (NDecoder decoder, t') -> (
+            match decoder.decode v with
             | Some v ->
                 matched_node := Some (t', D (decoder, v) :: decoded_values) ;
                 continue := false
             | None -> incr index )
-          | PLiteral lit, t' when String.equal lit uri_part ->
+          | `Path v, (NLiteral lit, t') when String.equal lit v ->
               matched_node := Some (t', decoded_values) ;
               continue := false
-          | PTrailing_slash, t' when String.equal "" uri_part ->
+          | `Path v, (NTrailing_slash, t') when String.equal "" v ->
               matched_node := Some (t', decoded_values) ;
               continue := false
-          | PFull_splat, t' ->
+          | `Path _, (NSplat, t') ->
               let path =
-                drop path_tokens matched_token_count |> String.concat "/"
+                drop path_tokens matched_token_count
+                |> List.map (fun (`Path tok) -> tok)
+                |> String.concat "/"
               in
               let splat_url =
                 String.split_on_char '?' uri
@@ -276,6 +316,16 @@ let rec match' method' uri (t : 'a t) =
               matched_node := Some (t', D (string, splat_url) :: decoded_values) ;
               continue := false ;
               full_splat_matched := true
+          | `Query (name, value), (NQuery_decoder (name', decoder), t') -> (
+            match decoder.decode value with
+            | Some v when String.equal name name' ->
+                matched_node := Some (t', D (decoder, v) :: decoded_values) ;
+                continue := false
+            | _ -> incr index )
+          | `Query (name1, value1), (NQuery_literal (name2, value2), t')
+            when String.equal name1 name2 && String.equal value1 value2 ->
+              matched_node := Some (t', decoded_values) ;
+              continue := false
           | _ -> incr index
         done ;
         Option.bind !matched_node (fun (t', decoded_values) ->
@@ -291,42 +341,45 @@ let rec match' method' uri (t : 'a t) =
       if i = n then None
       else
         match t.node_types.(i) with
-        | PMethod method'', t' when method_equal method' method'' ->
+        | NMethod method'', t' when method_equal method' method'' ->
             try_match t' [] uri_toks 0
         | _ -> (loop [@tailcall]) (i + 1)
     in
     loop 0
   else None
 
-and drop l n = match l with _ :: tl when n > 0 -> drop tl (n - 1) | t -> t
-
 and exec_route_handler : type a b. a -> (a, b) uri * decoded_value list -> b =
  fun f -> function
   | Nil, [] -> f
-  | Full_splat, [D (d, v)] -> (
+  | Splat, [D (d, v)] -> (
     match eq string.id d.id with Some Eq -> f v | None -> assert false )
   | Trailing_slash, [] -> f
-  | Method (_, uri), decoded_values -> exec_route_handler f (uri, decoded_values)
   | Literal (_, uri), decoded_values ->
       exec_route_handler f (uri, decoded_values)
-  | Decoder ({id; _}, uri), D ({id= id'; _}, v) :: decoded_values -> (
+  | Query_literal (_, _, uri), decoded_values ->
+      exec_route_handler f (uri, decoded_values)
+  | Decode ({id; _}, uri), D ({id= id'; _}, v) :: decoded_values -> (
+    match eq id id' with
+    | Some Eq -> exec_route_handler (f v) (uri, decoded_values)
+    | None -> assert false )
+  | Query_decode (_, {id; _}, uri), D ({id= id'; _}, v) :: decoded_values -> (
     match eq id id' with
     | Some Eq -> exec_route_handler (f v) (uri, decoded_values)
     | None -> assert false )
   | _, _ -> assert false
 
 module Private = struct
-  let route = route
   let nil = Nil
-  let trailing_slash = Trailing_slash
-  let full_splat = Full_splat
+  let splat = Splat
+  let t_slash = Trailing_slash
   let lit s uri = Literal (s, uri)
-  let decoder d uri = Decoder (d, uri)
-  let method' meth uri = Method (meth, uri)
+  let query_lit name value uri = Query_literal (name, value, uri)
+  let decode d uri = Decode (d, uri)
+  let query_decode name d uri = Query_decode (name, d, uri)
   let int = int
   let int32 = int32
   let int64 = int64
   let float = float
-  let bool = bool
   let string = string
+  let bool = bool
 end