+++ CHANGES.md
@@ -1,8 +1,13 @@
 ## v3.0.0 2021-08-02 
 
-- BREAKING CHANGE: A `full_splat` operator (`**`) now expects a last string paramter
-  in the handler. The parameter holds the remaining url matched by the
-`full_plat` operator
+- BREAKING CHANGE: A `splat` operator (`**`) now expects a last string paramter
+  in the handler. The parameter holds the remaining url matched by the `splat` operator
+- BREAKING CHANGE: 'create_decoder' has been renamed to 'decoder'
+- BREAKING CHANGE: 'create' has been renamed to 't'
+- NEW: Add `route` and `routes` functions
+- NEW: `%uri` ppx to create values of type `Wtr.uri`
+- FIX: Fix matching query components in uri
+- FIX: Fix pretty printing of query components in uri
 
 ## v2.0.0 2021-08-02